Modified the ScanProjectForSerializedTypes parameters to be more consistent in style with one another
Modified the ScanProjectForSerializedTypes parameters to be more consistent in style with one another by also having a scanResources parameter. Even though it is technically a little superfluous, it is a lot clearer in intent to also have a bool in style with the other parameters, instead of the user of the API having to infer that they must pass in an empty list as "resourcesToScan" to scan no resources at all.